fix(test): unblock company TEST initialization (#19)
* Merge remote-tracking branch 'origin/release' into feature

* fix(test): accept numeric zero-point catalog values

* chore(branch): sync release before migration MR

# Conflicts:
#	gaea/README.md
#	teacher/backend/OPERATIONS.md
#	teacher/backend/database/README.md

* feat(migrations): add staged company test upgrade

* fix(teacher): make personalized photo review ephemeral

* chore(branch): 同步 release 并解决迁移冲突 --task sso

* sso

* Merge remote-tracking branch 'origin/release' into HEAD

* feat(teacher): add personalized environment photo check

* Merge remote-tracking branch 'origin/release' into codex/g01-tesol-only-20260811

* Merge remote-tracking branch 'origin/feature' into codex/g01-tesol-only-20260811

# Conflicts:
#	teacher/backend/OPERATIONS.md
#	teacher/backend/src/platform/database/database.service.spec.ts
#	teacher/backend/src/platform/database/database.service.ts

* fix(g01): verify TESOL-only production readiness

* Merge remote-tracking branch 'origin/feature' into code

+++ deploy/combined/contract-probe.sql
@@ -383,7 +383,7 @@ BEGIN
           AND payload->>'completion_standard' =
               'The teacher app marks the task as completed after every requirement for the assigned improvement activity, including any required photo review, is satisfied.'
           AND payload->>'score_type' = 'ZERO'
-          AND (payload->>'score_value')::integer = 0
+          AND (payload->>'score_value')::numeric = 0
     ) THEN
         RAISE EXCEPTION
             'stable P-FB-NEGATIVE:v1 row is not the reviewed zero-point personalized improvement copy';
@@ -698,7 +698,7 @@ BEGIN
           AND template.status = 'PUBLISHED'
           AND template.execution_owner = 'TEACHER_APP'
           AND template.payload->>'category' = 'PERSONALIZED_IMPROVEMENT'
-          AND (template.payload->>'score_value')::integer = 0
+          AND (template.payload->>'score_value')::numeric = 0
     ) THEN
         RAISE EXCEPTION
             'P-FB-NEGATIVE is not the exact pending personalized photo execution';
